requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
torch
